requirements.txt
python-decouple==3.8
requests==2.32.3
terminaltables==3.1.10